refactor: reorganizing statement writers
