+++ src/modules/writers/customErrorWriter/implementations/CustomErrorWriter.ts
