+++ src/modules/writers/customErrorWriter/index.ts
