+++ src/modules/writers/customErrorWriter/types/ICustomErrorWriter.ts
