+++ src/modules/writers/globalMappingWriter/implementations/GlobalMappingWriter.ts
