+++ src/modules/writers/globalMappingWriter/index.ts
