+++ src/modules/writers/globalMappingWriter/test/globalMappingWriter.spec.ts
