+++ src/modules/writers/globalMappingWriter/types/IGlobalMappingWriter.ts
