+++ src/modules/writers/index.ts
@@ -4,4 +4,4 @@ import "./contractWriter";
 import "./eventWriter";
 import "./functionWriter";
 import "./globalVariableWriter";
-import "./variableWriter";
+import "./statements/variableWriter";
