+++ src/modules/writers/statements/index.ts
@@ -7,16 +7,16 @@ import "./inputWriter";
 import "./methodCallWriter";
 import "./newContractWriter";
 import "./outputWriter";
-import "./modifierWriter";
+import "../modifierWriter";
 import "./requireWriter";
 import "./assertWriter";
-import "./customErrorWriter";
+import "../customErrorWriter";
 import "./revertWriter";
 import "./breakWriter";
 import "./whileWriter";
 import "./doWhileWriter";
-import "./globalMappingWriter";
+import "../globalMappingWriter";
 import "./mappingWriter";
 import "./tryWriter";
 import "./catchWriter";
-import "./structWriter";
+import "../structWriter";
