+++ src/modules/writers/statements/variableWriter/implementations/VariableWriter.ts
