+++ src/modules/writers/statements/variableWriter/index.ts
