+++ src/modules/writers/statements/variableWriter/types/IVariableStatements.ts
