+++ src/modules/writers/statements/variableWriter/types/IVariableWriter.ts
