+++ src/modules/writers/structWriter/index.ts
