+++ src/modules/writers/structWriter/test/structWriter.spec.ts
