+++ src/modules/writers/structWriter/types/IStructWriter.ts
